backend/requirements.txt
openai
langchain
langchain-community
langchain-openai
faiss-cpu
chromadb
fastapi
uvicorn
python-dotenv